feat: add the ability to skip tests by runtime and flag #56
ISSUE: https://github.com/google/dotprompt/issues/56

RATIONALE:

Since we use YAML test suites to ensure that runtimes are at
feature parity, while we are working on some of these and some
runtimes are catching up, there are 2 ways in which to "catch up."

1. Maintain a list of working tests in the runtime code and
   add a way to identify the test. Over time the list would shrink
   as more features are implemented.

2. Add the ability to filter tests by runtime and flag to the
   test harness itself so that we can have a single source of truth
   indicating which tests work on which runtimes.

The first approach also requires adding a means to uniquely identify
a test and the second one appears to be a feature of the test harness
that would be useful later on if we have to deal with inconsistencies
between runtimes. I believe the latter approach is more declarative
and makes it easier to identify the tests that should be filtered
at the point of declaration.

EXAMPLE:

```
zsh❯ pnpm -C js run test

> dotprompt@1.0.1 test /Users/yesudeep/code/github.com/google/dotprompt/js
> vitest --run

The CJS build of Vite's Node API is deprecated. See https://vite.dev/guide/troubleshooting.html#vite-cjs-node-api-deprecated for more details.

 RUN  v3.0.5 /Users/yesudeep/code/github.com/google/dotprompt/js

 ✓ test/stores/dir.test.ts (18 tests) 265ms
 ✓ test/spec.test.ts (58 tests | 1 skipped) 57ms

 Test Files  2 passed (2)
      Tests  75 passed | 1 skipped (76)
   Start at  13:07:38
   Duration  927ms (transform 161ms, setup 0ms, collect 513ms, tests 322ms, environment 0ms, prepare 246ms)
```

CHANGELOG:
- [ ] Update the `spec.test.ts` test harness implementation to filter
  tests based on runtime support and an explicit flag.
- [ ] Add some tests to the harness itself to ensure that the filtering
  works as expected.
- [ ] Autoformat license headers in source.