feat: cargo workspace configuration and bazel build files for hermetic environment
RATIONALE:

Since we're supporting multiple languages in the repo, over time the
test running time and managing the environment to ensure we get hermetic
builds will become complicated. As a result, it is prudent to use a
proper build system to build and run tests orders of magnitude faster.

CHANGELOG:
- [ ] Add build files for go, ts.
- [ ] Add stub smoke tests for cargo workspace and java.
- [ ] Update setup script to install bazelisk.
- [ ] Add bazel configuration.
- [ ] Update hooks to run these tests.